pyproject.toml
[project]
name = "sodata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "vllm>=0.13.0",
]
